chore: Update android (WIP)
Android build was wildly out of date. This is the culmination of a few
days of battling gradle version issues, kotlin version issues, and java
version issues. In this commit, viro react compiles with each of the
dependencies (renderer, etc.) and creates the bridge properly both in a
React Native project and an Expo project.

There's a "hack" in the Expo withViroAndroid.ts file. Expo generates a
MainApplication.kt file that doesn't actually add the ViroReact packages
to the application.

Todo:
- Continue fixing arrow functions in some class components
- Rebuild virocore and debug crash in VROMathInvertMatrix
- Deploy release
- Add code to show a view instead of initializing AR if the device
  doesn't support AR Core

+++ prepare_release.sh
@@ -25,25 +25,26 @@ npm run test
 echo '========================================================================='
 echo 'Cleaning out the old build artifacts'
 echo '========================================================================='
-# rm android/viro_bridge/build/outputs/aar/*.aar
+rm android/viro_bridge/build/outputs/aar/*.aar
 
 echo '========================================================================='
 echo 'Building the React-Viro library'
 echo '========================================================================='
-# cd android && ./gradlew :viro_bridge:assembleRelease
+cd android && ./gradlew :viro_bridge:assembleRelease
+cd ..
 
 echo '========================================================================='
 echo 'Checking for build artifacts'
 echo '========================================================================='
-# if [ ! -f android/viro_bridge/build/outputs/aar/viro_bridge-release.aar ]; then
-#     echo -e "Unable to find viro-bridge release output!"
-#     exit
-# fi
+if [ ! -f android/viro_bridge/build/outputs/aar/viro_bridge-release.aar ]; then
+    echo -e "Unable to find viro-bridge release output!"
+    exit
+fi
 
 echo '========================================================================='
 echo 'Copying build artifacts to the lib directory'
 echo '========================================================================='
-# rm android/react_viro/*.aar
+rm android/react_viro/*.aar
 cp android/viro_bridge/build/outputs/aar/viro_bridge-release.aar android/react_viro/react_viro-release.aar
 
 echo '========================================================================='
